Use uint64_t to allow for large numbers of samples.
This solved issue with overlaps >1.

+++ src/Methods/SurfaceGeneration/include/VoxelCube.h
@@ -26,7 +26,7 @@
 
 class VoxelCube {
 public:
-    using VolumeDataType = uint32_t ;
+    using VolumeDataType = uint64_t ;
     using IndexType = dualmc::QuadIndexType;
     using VertexComponentsType = dualmc::VertexComponentsType;
